close: issue #120 (#123)
* feat : added step parameter into createXArray

* close: https://github.com/mljs/spectra-processing/issues/120

* chore: changed Uint16Array with DoubleArray to NumberArray. Returning a Float64Array when Float64Array returned.

* closed: https://github.com/mljs/spectra-processing/issues/119

* closed: https://github.com/mljs/spectra-processing/issues/118

* chore : updated the number array and removed "as number"

* chore: adding test cases for the step parameter

+++ src/x/xRolling.ts
@@ -50,11 +50,7 @@ export function xRolling(
 
   const newArray = [];
   for (let i = 0; i < array.length - window + 1; i++) {
-    let subArray = new Float64Array(
-      (array as Float64Array).buffer,
-      i * 8,
-      window,
-    );
+    let subArray = new Float64Array(array.buffer, i * 8, window);
     // we will send a view to the original buffer
     newArray.push(fct(subArray));
   }
